moves migrate command around in test helper

+++ test/test_helper.rb
@@ -3,10 +3,11 @@ require 'rubygems'
 require 'minitest/autorun'
 require 'rack/test'
 
-File.delete('test/stations_bot-test.db')
+require 'sequel'
+File.delete('test/stations_bot-test.db') if File.exist?('test/stations_bot-test.rb')
+require File.expand_path('../../db/connect.rb', __FILE__)
+Sequel.extension :migration
+Sequel::Migrator.run(DB, 'db/migrations')
 
 require File.expand_path('../../bootstrap.rb', __FILE__)
 require File.expand_path('../stations_bot_test_base.rb', __FILE__)
-
-Sequel.extension :migration
-Sequel::Migrator.run(DB, 'db/migrations')
